Tuning and correcting a bit wrt. stats timings.
Can only give timing on the node that performed the full operation, i.e.
from the first stats request, to the last reply gotten. The request must
have been performed on "this node" to be able to calculate the
difference (or, we could have done it on both nodes, but thus risking
time skews on the other nodes).