Bit of refactors and JavaDoc before adding new features.
Added more JavaDoc to API.

Changed return values of the delete and reissue methods to return a map
of msgSysId->MatsMessageMetadata so as to be able to include more than
one piece of info, and also future proof it more.

Impl detail: HTML/JavaScript GUI: Renamed "_bulk" operations to
"_selected".