Refactoring MatsBrokerMonitor to handle multiple queues per stage
.. outside of the standard incoming and DLQ.

It is now an Enum instead of 2x all methods. The currently defined are:
1. STANDARD
2. DEAD_LETTER_QUEUE
3. NON_PERSISTENT_INTERACTIVE
4. DEAD_LETTER_QUEUE_NON_PERSISTENT_INTERACTIVE
5. MUTED_DEAD_LETTER_QUEUE
6. WIRETAP

None of them are currently used - and in this commit, they are not shown
in the GUI either. However, the end result should currently be identical
to the previous functionality.

The reason for doing this is in first round for handling "Muted DLQs" -
that is DLQs that shouldn't anymore be warned on in a monitoring system,
because they're already being investigated.