Improving the multi-queue solution some more
Mainly: Do not show all the queues that have 0 messages in the
BrokerOverview, instead just tallying "+2q" if there are two more queues
with zero messages. Instead, in the BrowseQueue view, all queues will be
shown top right (effectively same view as in the BrokerOverview stage
thingy before) - so that you can see them, and go to them.